arm64: dts: remove alloc-ranges property in cma nodes for tl1 [1/1]
PD#SWPL-3437

Problem:
the cam regions that specified by the alloc-ranges property are scattered,
it is not convenient to debug the memory issues.

Solution:
remove alloc-ranges property in cam nodes, but secmon and fb. the system
will allocate the cma region automatically that more gathered.

Verify:
test pass on x301

Change-Id: Ibbe64b48843c1295110242c17d7691284f08a66b
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>