wdt: add option to avoid the wdt is disabled casually [1/1]
PD#SWPL-14925

Problem:
The system can't be reset by wdt which is fed by the tv-service
when the tv-service is abnormally hanged.

Solution:
Add extra option to avoid the wdt is disabled when the fed dog
thread aborts, but we can disable wdt when certain condition is
met.

Verify:
test pass on customer board

Change-Id: I1fd5ff4e36f06bda99726c0a2ced399299596f95
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>