sdio: optimize sdio suspend/resume [1/1]
PD#SWPL-16036

Problem:
When the system starts up, wifi is turned on.
When we manually turn off wifi and start the
suspend & resume operation, we will find that
the sdio error is reported when we wake up

Solution:
Add the sdio_notify() interface to notify the
sdio controller when wifi is on/off, enabling
the keep_power standard to be configured when suspend.

Verify:
TL1

Change-Id: I48e0c31bf7ea80d3efd4b0d41b4af378e859dc15
Signed-off-by: Nan Li <nan.li@amlogic.com>