hdmitx:  HDMI HDCP didn't work in DRM mode [1/1]
PD#SWPL-17207

Problem:
The variable of hdcp_tx_type isn't updated before the verification
of HDCP22.
The time is so short that loading the process of "hdcp_tx22" isn't
in time.
The Hpd state and mode set flag isn't updated to upper layer.
If the verification of hdcp22 is fail, there is no code responsible
to check Hpd state and mode set flag.

Solution:
Update hdcp_tx_type before the verification of HDCP22.
Modify timeout of the vevification of HDCP22 from 10s to 50s.
Update the Hpd state and mode set flag to upper layer.
If the verification of hdcp22 is fail, check Hpd state and
mode set flag.

Verify:
verify on the board of U212

Change-Id: I284708545cf9a6c9bd10cac70c9c0ad8768f6f05
Signed-off-by: Zhengrong Zhu <zhengrong.zhu@amlogic.com>
Signed-off-by: Zhengrong.Zhu <zhengrong.zhu@amlogic.com>
Signed-off-by: Zhengrong Zhu <zhengrong.zhu@amlogic.com>