hdmitx: Four Block Edid Read Error [1/1]
PD#SWPL-28693

Problem:
In the file of hdmi_tx_ddc.c, the registers of
HDMITX_DWC_I2CM_SEGADDR and HDMITX_DWC_I2CM_SEGPTR and set to 0x30
and 0x01, but don't clear these registers.
In the file of drm_edid.c of DRM framework，the process of read
isn't suitable to our chip. The process of read is to write slave
addr and then restart for read. But the process of read is done by
hw in our chip. We config the registers only.
In the file of meson_hdmi.c, the register of HDMITX_DWC_I2CM_SLAVE
is set to DDC_SEGMENT_ADDR(0x30) errorly and isn't set to 0x50, when
read block2/3.

Solution:
In the file of hdmi_tx_ddc.c, the registers of
HDMITX_DWC_I2CM_SEGADDR and HDMITX_DWC_I2CM_SEGPTR is set to zero.
In the file of drm_edid.c of DRM framework，delete the process of
write slave addr for reading.
In the file of meson_hdmi.c, the register of HDMITX_DWC_I2CM_SLAVE
is set to 0x50, read any block.

Verify:
Verify on the board of U212

Change-Id: I967534e18136540d8b62c4646758c0b2b6f4493a
Signed-off-by: Zhengrong Zhu <zhengrong.zhu@amlogic.com>