remote: power key event is repeatedly reported when the system wakes up [1/1]
PD#SWPL-30120

Problem:
If we choose the legacy IR controller to respone the IR-KEY after system
standby, an invalid interrupt was triggered when the system wakes up, and
the interrupt will cause the power key event is repeatedly reported at the
wakeup phase.

Solution:
If the data frame is invalid, the interrupt handle should be ignored.

Verify:
test pass on txlx

Change-Id: If3a57d74e9d3564944d1a4f80975e8754facaecd
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>