vout: remove aml_tvout_mode_monitor [1/1]
PD#SWPL-32713

Problem:
Plug in the HDMI cable when kernel is just booting, the tvout
monitor will check and set the hdmi mode, but it may set the
wrong mode, such as 2160p60hz RGB 12bit

Solution:
When kernel is just booting, all things are preparing and not
real ready to use. And hdmi mode setting should be governed by
the systemcontrol, plus the OSD/VPP/DV, ... etc.

Verify:
SM1

Change-Id: Id3e9a21d4a0c1156cf4b10ecf9afa20a1a358ccc
Signed-off-by: Zongdong Jiao <zongdong.jiao@amlogic.com>