audio: fix spdif open release samesrc [1/1]
PD#SWPL-58748

Problem:
1. TDM same source open and write
2. SPDIF then open and write
As the tdm samesrc is still there but spdif output
is not ready to prepare(buffer level less than start threshold),
then the tdm output HW is waiting for spdif to start, but the spdif
writing thread(same writing thread to tdm out) is waiting for available
buffer to write. This comes to deadlock.

Solution:
Add release samesrc when spdif open if the tdm same source is on.

Verify:
SC2

Signed-off-by: Shuai Li <shuai.li@amlogic.com>
Change-Id: I0405c5b126c52ce7b8d7cafc99511b4e2885d477