ext4: fix race writing to an inline_data file while its xattrs are changing (CVE-2021-40490)
PD#SWPL-75272

Notice: this object is not reachable from any branch.
The location of the system.data extended attribute can change whenever
xattr_sem is not taken.  So we need to recalculate the i_inline_off
field since it mgiht have changed between ext4_write_begin() and
ext4_write_end().

This means that caching i_inline_off is probably not helpful, so in
the long run we should probably get rid of it and shrink the in-memory
ext4 inode slightly, but let's fix the race the simple way for now.

Cc: stable@kernel.org
Fixes: f19d5870cbf72 ("ext4: add normal write support for inline data")
Reported-by: syzbot+13146364637c7363a7de@syzkaller.appspotmail.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Notice: this object is not reachable from any branch.

Change-Id: Iab4c5d8d3e231d036858c0d5c3ced756fa5f9809
Signed-off-by: yao zhang1 <yao.zhang1@amlogic.com>