ANDROID: dm-bow: serialize state changes with completion of writes
dm-bow has a long-standing race condition. When in the CHECKPOINT state,
writes are completed asynchronously on a workqueue, rather than
occurring synchronously within dm_bow_map() as they are when in the
COMMITTED state. There is no synchronization between the completion of
these writes and the state change, leading to potential corruption.

Specifically, problems occur during the transition from CHECKPOINT to
COMMITTED. When state_store() triggers this state change, it sets an
atomic variable that causes all subsequent writes to follow the
synchronous write path used by the COMMITTED state. Any already
in-flight writes on the workqueue complete asynchronously, however, and
as if still in CHECKPOINT. When bow_write() runs for these writes, any
blocks that were considered free while in CHECKPOINT are considered safe
to write to. Some of those blocks, however, may have already been
recently written to via the synchronous COMMITTED path.

When that happens, those recently-written post-commit blocks will be
overwritten by the pre-commit write. This results in both corruption of
the blocks written after the state transitioned to COMMITTED as well as
the loss of the writes that were still in-flight during the state
transition.

Fix this by adding the missing synchronization. Acquire a mutex when
changing the state and, with this held, flush the workqueue to ensure
that all writes prior to the state change have complete. Use this same
lock to ensure no new writes will be performed/queued until the state
change is complete. This mutex adds negligible overhead since the only
hot path it is acquired on is one that's already serialized via the
existing ranges_lock.

Fixes: 0ce3eb37e9ea ("ANDROID: dm-bow: Add dm-bow feature")
Change-Id: I5e7e11d6a43ab7eb9fe95b8c4065def564608fcd
Signed-off-by: Matt Wagantall <mattnw@meta.com>
Signed-off-by: Jiucheng Xu <jiucheng.xu@amlogic.com>